Polish important rules

+++ fix/rules.js
@@ -1,9 +1,12 @@
 import * as fs from 'node:fs/promises'
+import * as util from 'node:util'
 import detectIndent from 'detect-indent'
 import { execa } from 'execa'
-import * as util from 'node:util'
 import * as _ from 'lodash-es'
+import dedent from 'dedent'
 import { fileExists } from './util.js'
+import * as diff from 'diff'
+import chalk from 'chalk'
 
 /**
  * @typedef {import('../index.js')} Rule
@@ -22,15 +25,24 @@ export async function* filesMustHaveContent(mapping) {
 		}
 
 		if (expectedContent === null) {
-			yield {
-				title: `File "${file}" must not exist`,
-				fix: () => fs.rm(file)
+			if (await fileExists(file)) {
+				yield {
+					message: [`Expected file "${file}" to not exist`, 'But, found the file'],
+					fix: () => fs.rm(file)
+				}
 			}
 		} else {
-			const content = await fs.readFile(file, 'utf-8')
-			if (content !== expectedContent) {
+			if (await fileExists(file)) {
+				const content = await fs.readFile(file, 'utf-8')
+				if (content !== expectedContent) {
+					yield {
+						message: `  => Expected file "${file}" to have content:\n---\n${expectedContent}\n---\n  => But, the file has content:\n---\n${content}\n---\n`,
+						fix: () => fs.writeFile(file, expectedContent)
+					}
+				}
+			} else {
 				yield {
-					title: `File "${file}" must have content: "${expectedContent}"`,
+					message: `  => Expected file "${file}" to exist and have content:\n---\n${expectedContent}\n---\n  => But, the file does not exist`,
 					fix: () => fs.writeFile(file, expectedContent)
 				}
 			}
@@ -44,155 +56,90 @@ export async function* filesMustHaveContent(mapping) {
  */
 export async function* filesMustHaveShape(mapping) {
 	for (let file in mapping) {
-		const shape = mapping[file]
+		const source = mapping[file]
 
 		if (file.slice(0, 2) === './') {
 			file = file.slice(2)
 		}
 
-		/**
-		 * @param {Record<string, Record<string, unknown>>} object
-		 * @param {Record<string, unknown>} shape
-		 * @returns {Record<string, unknown>}
-		 */
-		function merge(object, shape) {
-			for (const [propChain, propValue] of Object.entries(shape)) {
-				if (Object.hasOwn(propValue, '__delete')) {
-					object = _.omit(object, propChain)
-				} else if (Object.hasOwn(propValue, '__replace')) {
-					_.set(object, propChain, propValue)
+		const content = await fs.readFile(file, 'utf-8')
+		const actual = JSON.parse(content)
+		const expected = structuredClone(actual)
+		customMerge(expected, source)
+
+
+		if (!util.isDeepStrictEqual(expected, actual)) {
+			let difference = ''
+			for (const part of diff.diffJson(JSON.stringify(actual, null, 2), JSON.stringify(expected, null, 2))) {
+				if (part.added) {
+					difference += chalk.green(part.value)
+				} else if (part.removed) {
+					difference += chalk.red(part.value)
 				} else {
-					// __merge, the default.
-					const obj = _.get(object, propChain)
-					if (obj === undefined) {
-						_.set(object, propChain, propValue)
-					} else {
-						_.merge(obj, propValue)
-					}
+					difference += part.value
 				}
 			}
 
-			return object
-		}
-
-		const content = await fs.readFile(file, 'utf-8')
-		const actualObject = JSON.parse(content)
-		const expectedObject = merge(structuredClone(actualObject), shape)
-
-		yield {
-			async issues() {
-				return !util.isDeepStrictEqual(expectedObject, actualObject)
-			},
-			printInfo() {
-				console.log(`File "${file}" must have the correct shape`)
-			},
-			async fix() {
-				await fs.writeFile(
+			yield {
+				message: '  ' + dedent`
+					=> Expected file "${file}" to have the correct shape:
+					---
+					${difference.replaceAll('\n', '\n' + '\t'.repeat(5))}
+					---
+					`,
+				fix: () => fs.writeFile(
 					file,
-					JSON.stringify(expectedObject, null, detectIndent(content).indent || '\t'),
+					JSON.stringify(expected, null, detectIndent(content).indent || '\t'),
 				)
-			},
-		}
-	}
-}
-
-/**
- * @typedef ruleCheckPackageJsonDependenciesParam
- * @property {string[]} packages
- *
- * @param {ruleCheckPackageJsonDependenciesParam} param0
- * @returns {AsyncGenerator<Rule>}
- */
-export async function* ruleCheckPackageJsonDependencies({ packages }) {
-	const packageJsonText = await fs.readFile('package.json', 'utf-8')
-	/** @type {import('type-fest').PackageJson} */
-	const packageJson = JSON.parse(packageJsonText)
-
-	const latestVersionsObjects = await Promise.all(
-		packages.map((packageName) => execa('npm', ['view', '--json', packageName])),
-	)
-	const latestVersions = latestVersionsObjects.map((result) => {
-		if (result.exitCode !== 0) {
-			console.log(result.stderr)
-			throw new Error(result)
-		}
-
-		const obj = JSON.parse(result.stdout)
-		return obj['dist-tags'].latest
-	})
-
-	let shouldFix = false
-	for (const packageName of packages) {
-		if (!packageJson?.devDependencies?.[packageName]) {
-			shouldFix = true
-		}
-	}
-
-	for (let i = 0; i < packages.length; ++i) {
-		const packageName = packages[i]
-		// TODO: ^, etc. is not always guaranteed
-		if (packageJson?.devDependencies?.[packageName].slice(1) !== latestVersions[i]) {
-			shouldFix = true
-		}
-	}
-
-
-	if (shouldFix) {
-		yield {
-			title: `File 'package.json' is missing dependencies for package`,
-			fix
-		}
-	}
-
-	async function fix() {
-		const packageJsonModified = structuredClone(packageJson)
-		for (let i = 0; i < packages.length; ++i) {
-			const packageName = packages[i]
-
-			// TODO: ^, etc. should not always be done
-			packageJsonModified.devDependencies = {
-				...packageJsonModified?.devDependencies,
-				[packageName]: `^${latestVersions[i]}`,
 			}
 		}
-
-		await fs.writeFile(
-			'package.json',
-			JSON.stringify(packageJsonModified, null, detectIndent(packageJsonText).indent),
-		)
-		console.log(`Now, run: 'npm i`)
 	}
 }
 
 /**
- * @typedef ruleCheckPackageJsonDependenciesParam2
- * @property {string} mainPackageName
- * @property {Record<string, string | null>} packages
- *
- * @param {ruleCheckPackageJsonDependenciesParam2} param0
- * @returns {AsyncGenerator<Rule>}
+ * @param {Record<string, any>} target
+ * @param {Record<string, any>} source
  */
-export async function* ruleCheckPackageJsonDependencies2({ mainPackageName, packages }) {
-	const packageJsonText = await fs.readFile('package.json', 'utf-8')
-	/** @type {import('type-fest').PackageJson} */
-	const packageJson = JSON.parse(packageJsonText)
-
-	let shouldFix = false
-	for (const [packageName, packageValue] of Object.entries(packages)) {
-		if (!packageJson?.devDependencies?.[packageName] && packageValue !== null) {
-			shouldFix = true
-		}
-	}
-
-	for (const [packageName, packageValue] of Object.entries(packages)) {
-		if (packageJson?.devDependencies?.[packageName] && packageValue === null) {
-			shouldFix = true
-		}
-	}
-
-	if (shouldFix) {
-		yield {
-			title: `File 'package.json' is missing dependencies2 for package: ${mainPackageName}`
+export function customMerge(target, source) {
+	for (const key in source) {
+		if (Object.hasOwn(target, key)) {
+			if (typeof source[key] === 'object' && source[key] !== null) {
+				if (Object.hasOwn(source[key], '__delete')) {
+					delete target[key]
+				} else if (Object.hasOwn(source[key], '__replace')) {
+					delete source[key].__replace
+					target[key] = source[key]
+					customMerge(target[key], source[key])
+				} else {
+					Object.assign(target[key], source[key])
+					customMerge(target[key], source[key])
+				}
+			} else {
+				target[key] = source[key]
+				if (typeof target[key] === 'object' && target[key] !== null) {
+					customMerge(target[key], source[key])
+				}
+			}
+		} else {
+			if (typeof source[key] === 'object' && source[key] !== null) {
+				if (Object.hasOwn(source[key], '__delete')) {
+				} else if (Object.hasOwn(source[key], '__replace')) {
+					target[key] = source[key]
+					customMerge(target[key], source[key])
+				} else {
+					target[key] = {}
+					Object.assign(target[key], source[key])
+					customMerge(target[key], source[key])
+					if (Object.keys(target[key]).length === 0) {
+						delete target[key]
+					}
+				}
+			} else {
+				target[key] = source[key]
+				if (typeof target[key] === 'object' && target[key] !== null) {
+					customMerge(target[key], source[key])
+				}
+			}
 		}
 	}
 }
